// src/Envirotrax.App/envirotrax.app.client/src/app/shared/models/public-search/public-csi-inspection-result.ts
export interface PublicCsiInspectionResult {
    id: number;
    inspectionDate?: string;
    propertyBusinessName?: string;
    propertyStreetNumber?: string;
    propertyStreetName?: string;
    propertyNumber?: string;
    propertyCity?: string;
    propertyState?: string;
    propertyZip?: string;
    inspectorCompanyName?: string;
    inspectorContactName?: string;
    inspectorAddress?: string;
    inspectorCity?: string;
    inspectorState?: string;
    inspectorZip?: string;
}